refactor: restructure project — rename files to remove redundant prefixes, move shared components to sections/, and rename services-data to services
