+++ components/cosmos/background.tsx
