--- components/layout/page-wrapper.tsx
@@ -1,5 +1,0 @@
-import type { ReactNode } from 'react'
-
-export default function PageWrapper({ children }: { children: ReactNode }) {
-  return <div className='cs-page cs-fade-in'>{children}</div>
-}
